requirements.txt
streamlit
streamlit.components.v1
numpy
pandas
scipy 
scipy.interpolate
bokeh.models 
bokeh.plotting
pynumpress
pyteomics
requests 
io 
plotly.graph_objects
py3Dmol

pydeck
